Revert "change: ホームページを作成"
This reverts commit 787a95a7dbb41a0d772c8ed999a7dfa5b83981f9.
