+++ app/layout.tsx
@@ -1,6 +1,5 @@
 import type { Metadata } from 'next';
 import './globals.css';
-import { Header } from './layout/header';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
@@ -14,11 +13,10 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className='w-screen h-screen flex flex-col gap-2 items-center sm:items-start'>
-        <Header />
-        <main className='w-full h-full'>
-          {children}
-        </main>
+      <body>
+        <div className="w-screen h-screen">
+          <div className="w-full h-full">{children}</div>
+        </div>
       </body>
     </html>
   );
